Fix PrioritySelectedProjectsRepository resource retrieval to follow first-seen policy
The get_resource method was incorrectly searching across all repositories for a resource, but should follow the same first-seen policy as get_project_page.

In the wild, I saw this when we released a wheel internally with the same name as one on pypi. Our one didn't include a metadata file, and when we went to fetch the metadata it was being taken from the pypi one. The result was incorrect dependendencies on install.